refactor(instrumentation): removed shared folder from instrumentation path
