+++ app/gateways/db/postgres/get_account_balance.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/stone-co/the-amazing-ledger/app"
 	"github.com/stone-co/the-amazing-ledger/app/domain/vos"
-	"github.com/stone-co/the-amazing-ledger/app/shared/instrumentation/newrelic"
+	"github.com/stone-co/the-amazing-ledger/app/instrumentation/newrelic"
 )
 
 const getAccountBalanceQuery = `
