+++ app/gateways/db/postgres/get_analytical_data.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/stone-co/the-amazing-ledger/app/domain/vos"
-	"github.com/stone-co/the-amazing-ledger/app/shared/instrumentation/newrelic"
+	"github.com/stone-co/the-amazing-ledger/app/instrumentation/newrelic"
 )
 
 const analyticalDataQuery = `
